+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 1.19.4-3.0.6
+## 1.19.4-3.1.0
+
 **Added**
-- Compatibility for Inventorio has been added back (as Inventorio is now updated to 1.19.4).
+
+- Graves are now backed up on the server, accessible even if the grave has already been claimed.
+
+- Added four new commands to interact with grave backups: `players`, `list`, `restore`, and `delete`. You can read about them [here](https://github.com/ginsm/forgotten-graves/wiki/Commands#grave-backup-commands).
+
+- Added two new config settings for grave backups: `storedGravesAmount` and `destructiveDeleteCommand`. You can read about them [here](https://github.com/ginsm/forgotten-graves/wiki/Config#stored-graves-amount).